refactor(controller/kube): move DinD const/size seam to render_dind.go (1/3)
